requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
supabase==2.9.1
gspread==6.1.4
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.116.0
aiohttp==3.10.11
python-multipart==0.0.12
python-dateutil==2.9.0
pydantic==2.9.2
typing-extensions==4.12.2
python-dotenv==1.0.0
Flask==3.0.0
requests==2.31.0
httpx[socks]==0.26.0
